move number of recipes card to components and callbacks
